feat: change from Exeption Pattern to Result Pattern (not check error)

+++ Application/Features/Outputs/Commands/CreateOutputByManager/CreateOutputProductByManagerCommandValidator.cs
@@ -9,7 +9,7 @@ namespace Application.Features.Outputs.Commands.CreateOutputByManager
         {
             RuleFor(x => x.ProductId).NotNull().GreaterThan(0);
 
-            RuleFor(x => x.Count).NotNull().GreaterThan((short)0);
+            RuleFor(x => x.Count).NotNull().GreaterThan(0);
         }
     }
 }
